Added error compensation For lift PID
Update counts when buttons are not being pressed as well as being pressed.
Added error compensation to LiftSystemPID setpoint given gravity's affect on the downward position and the momentums alteration of position when the system is lifted. This compensation ensures that the PID is not attempting to reach a point before the lift actually stops. The encoder count for the lift is not exact and so values must be added to the setpoint depending on whether the lift was brought up or down previously. Current testing shows this to work with no slippage or jolting. Further drive testing will either present more errors or prove the efficacy of these measures.