Ciommented out Grip Vision Code
The Radio seemed to be using all the bandwidth on the camera. The radio would be connected to the computer and video feed would be fine; however, the the robot would not connect through the driver station. After commenting all the vision code out and closing Grip, this issue ceased.